requirements.txt
astor==0.8.1
brotlipy==0.7.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
decorator==5.1.1
docopt==0.6.2
easydict==1.10
idna==3.4
imageio==2.22.4
joblib==1.2.0
numpy==1.23.5
opt-einsum==3.3.0
paddle-bfloat==0.1.7
paddlepaddle-gpu==2.4.1.post112
Pillow==9.2.0
pip==22.3.1
pipreqs==0.4.11
protobuf==3.20.0
pycparser==2.21
pyOpenSSL==22.1.0
PySocks==1.7.1
PyYAML==6.0
requests==2.28.1
scikit-learn==1.2.0
scipy==1.9.3
setuptools==65.5.1
six==1.16.0
threadpoolctl==3.1.0
tqdm==4.64.1
urllib3==1.26.13
wheel==0.38.4
yarg==0.1.9
